chore(ci) refactor hybrid e2e tests to speed them up
Don't recreate CP and apps for every tests as we're only asserting that things are
the way we want
