+++ test/e2e/hybrid/globalkubernetes/kuma_hybrid_kube_global.go
@@ -1,4 +1,4 @@
-package hybrid
+package globalkubernetes
 
 import (
 	. "github.com/onsi/ginkgo"
