+++ test/e2e/hybrid/globaluniversal/e2e_suite_test.go
@@ -1,9 +1,12 @@
-package hybrid_test
+package globaluniversal_test
 
 import (
 	"testing"
 
+	. "github.com/onsi/ginkgo"
+
 	"github.com/kumahq/kuma/pkg/test"
+	"github.com/kumahq/kuma/test/e2e/hybrid/globaluniversal"
 	"github.com/kumahq/kuma/test/framework"
 )
 
@@ -14,3 +17,5 @@ func TestE2EDeploy(t *testing.T) {
 		t.SkipNow()
 	}
 }
+
+var _ = Describe("Test Kubernetes/Universal deployment", globaluniversal.KubernetesUniversalDeployment)
