--- test/e2e/hybrid/kuma_hybrid_kube_global_test.go
@@ -1,9 +1,0 @@
-package hybrid_test
-
-import (
-	. "github.com/onsi/ginkgo"
-
-	"github.com/kumahq/kuma/test/e2e/hybrid"
-)
-
-var _ = Describe("Test Kubernetes/Universal deployment when Global is on K8S", hybrid.KubernetesUniversalDeploymentWhenGlobalIsOnK8S)
